Move OpenSSL includes and config to separate file
Also move some of the common functionality that's used in the NIF
implementations.
